build: get Makefile working with src & test files
So, so many changes. Basically rewrote the entire Makefile using static pattern rules, and a special rule that reads a var set when calling the command (make test file=./path/to/file)

Renaming the file too because GNU make will look for GNUmakefile, makefile, or Makefile. https://www.gnu.org/software/make/manual/html_node/Makefile-Names.html

Quite annoying to set up and get right, but it should be flexible and adapt as more files are added, as it is based on file name patterns